backend/requirements.txt
fastapi==0.124.0

uvicorn==0.38.0
motor==3.7.1
pymongo>=4.6.0
python-jose==3.3.0

passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.2.1
pydantic==2.12.5
python-multipart==0.0.20
email-validator==2.3.0
pydantic-settings==2.10.1
